refactor: reduce rerenders in TradeConfirmation
